+++ apps/frontend/assets/src/omg-page-blockchain-wallets/BlockchainSettingsPage.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const BlockchainSettingsPage = () => {
+  return <div>BlockchainSettingsPage</div>
+}
+
+export default BlockchainSettingsPage
