+++ apps/frontend/assets/src/omg-page-blockchain-wallets/BlockchainTokensPage.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const BlockchainTokensPage = () => {
+  return <div>BlockchainTokensPage</div>
+}
+
+export default BlockchainTokensPage
