+++ apps/frontend/assets/src/omg-page-blockchain-wallets/BlockchainTransactionsPage.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const BlockchainTransactionsPage = () => {
+  return <div>BlockchainTransactionsPage</div>
+}
+
+export default BlockchainTransactionsPage
