+++ apps/frontend/assets/src/omg-page-blockchain-wallets/InternalTokensPage.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const InternalTokensPage = () => {
+  return <div>InternalTokensPage</div>
+}
+
+export default InternalTokensPage
